requirements.txt
alembic~=1.6.5
psycopg2-binary~=2.9.1
Flask~=2.0.1
SQLAlchemy-serializer~=1.4.1
python-dotenv~=0.19.0
requests~=2.26.0